+++ webferea/__init__.py
@@ -49,7 +49,7 @@ def create_app(test_config=None):
     from . import feed
     from . import entry
     from . import sync
-    from .helpers import format_datetime, format_iframes, format_internal_links
+    from .helpers import format_datetime, format_iframes, format_internal_links, filter_bleach
 
     db.init_app(app)
     app.register_blueprint(auth.bp)
@@ -60,5 +60,6 @@ def create_app(test_config=None):
     app.jinja_env.filters['datetime'] = format_datetime
     app.jinja_env.filters['iframes'] = format_iframes
     app.jinja_env.filters['internal_links'] = format_internal_links
+    app.jinja_env.filters['bleach'] = filter_bleach
 
     return app